refactor: split internal functionality from workflow service
